Some fixes, touched the boss a bit, still needs work!

+++ Assets/Scripts/WorldHealthBar.cs
@@ -12,6 +12,9 @@ public class WorldHealthBar : MonoBehaviour
 		if (target == null) return;
 
 		transform.position = target.transform.position + offset;
-		fillImage.fillAmount = target.Health.CurrentHP / target.maxHP;
+		// Cast to float before dividing. Without the cast, C# performs integer
+		// division (e.g. 75 / 100 = 0), which made the health bar snap instantly
+		// between full and empty with no intermediate states.
+		fillImage.fillAmount = (float)target.Health.CurrentHP / target.maxHP;
 	}
 }
